helper functions added
