+++ packages/headless/src/features/result-actions/result-actions-insight-analytics-actions.ts
@@ -1,12 +1,14 @@
 import {InsightPanel} from '@coveo/relay-event-types';
 import {Result} from '../../api/search/search/result';
 import {
+  analyticsEventItemMetadata,
   documentIdentifier,
   InsightAction,
   makeInsightAnalyticsAction,
   partialDocumentInformation,
   validateResultPayload,
 } from '../analytics/analytics-utils';
+import {analyticsEventCaseContext} from '../analytics/insight-analytics-utils';
 import {getCaseContextAnalyticsMetadata} from '../case-context/case-context-state';
 
 export const logCopyToClipboard = (result: Result): InsightAction =>
@@ -25,27 +27,13 @@ export const logCopyToClipboard = (result: Result): InsightAction =>
     },
     analyticsType: 'InsightPanel.ItemAction',
     analyticsPayloadBuilder: (state): InsightPanel.ItemAction => {
-      const metadata = getCaseContextAnalyticsMetadata(
-        state.insightCaseContext
-      );
-      const identifier = documentIdentifier(result);
       const information = partialDocumentInformation(result, state);
       return {
-        itemMetadata: {
-          uniqueFieldName: identifier.contentIDKey,
-          uniqueFieldValue: identifier.contentIDValue,
-          title: information.documentTitle,
-          author: information.documentAuthor,
-          url: information.documentUri,
-        },
+        itemMetadata: analyticsEventItemMetadata(result, state),
         position: information.documentPosition,
         searchUid: state.search?.response.searchUid || '',
         action: 'copyToClipboard',
-        context: {
-          targetId: metadata.caseId || '',
-          targetType: 'Case',
-          caseNumber: metadata.caseNumber,
-        } as InsightPanel.Context,
+        context: analyticsEventCaseContext(state),
       };
     },
   });
@@ -66,27 +54,13 @@ export const logCaseSendEmail = (result: Result): InsightAction =>
     },
     analyticsType: 'InsightPanel.ItemAction',
     analyticsPayloadBuilder: (state): InsightPanel.ItemAction => {
-      const metadata = getCaseContextAnalyticsMetadata(
-        state.insightCaseContext
-      );
-      const identifier = documentIdentifier(result);
       const information = partialDocumentInformation(result, state);
       return {
-        itemMetadata: {
-          uniqueFieldName: identifier.contentIDKey,
-          uniqueFieldValue: identifier.contentIDValue,
-          title: information.documentTitle,
-          author: information.documentAuthor,
-          url: information.documentUri,
-        },
+        itemMetadata: analyticsEventItemMetadata(result, state),
         position: information.documentPosition,
         searchUid: state.search?.response.searchUid || '',
         action: 'sendEmail',
-        context: {
-          targetId: metadata.caseId || '',
-          targetType: 'Case',
-          caseNumber: metadata.caseNumber,
-        } as InsightPanel.Context,
+        context: analyticsEventCaseContext(state),
       };
     },
   });
@@ -107,27 +81,13 @@ export const logFeedItemTextPost = (result: Result): InsightAction =>
     },
     analyticsType: 'InsightPanel.ItemAction',
     analyticsPayloadBuilder: (state): InsightPanel.ItemAction => {
-      const metadata = getCaseContextAnalyticsMetadata(
-        state.insightCaseContext
-      );
-      const identifier = documentIdentifier(result);
       const information = partialDocumentInformation(result, state);
       return {
-        itemMetadata: {
-          uniqueFieldName: identifier.contentIDKey,
-          uniqueFieldValue: identifier.contentIDValue,
-          title: information.documentTitle,
-          author: information.documentAuthor,
-          url: information.documentUri,
-        },
+        itemMetadata: analyticsEventItemMetadata(result, state),
         position: information.documentPosition,
         searchUid: state.search?.response.searchUid || '',
         action: 'postToFeed',
-        context: {
-          targetId: metadata.caseId || '',
-          targetType: 'Case',
-          caseNumber: metadata.caseNumber,
-        } as InsightPanel.Context,
+        context: analyticsEventCaseContext(state),
       };
     },
   });
